+++ node/src/cli.rs
@@ -1,16 +1,16 @@
 use crate::chain_spec;
+use clap::Parser;
 use std::path::PathBuf;
-use structopt::StructOpt;
 
 /// Sub-commands supported by the collator.
-#[derive(Debug, StructOpt)]
+#[derive(Debug, clap::Subcommand)]
 pub enum Subcommand {
     /// Export the genesis state of the parachain.
-    #[structopt(name = "export-genesis-state")]
+    #[clap(name = "export-genesis-state")]
     ExportGenesisState(ExportGenesisStateCommand),
 
     /// Export the genesis wasm of the parachain.
-    #[structopt(name = "export-genesis-wasm")]
+    #[clap(name = "export-genesis-wasm")]
     ExportGenesisWasm(ExportGenesisWasmCommand),
 
     /// Build a chain specification.
@@ -35,7 +35,7 @@ pub enum Subcommand {
     Revert(sc_cli::RevertCmd),
 
     /// The custom benchmark subcommmand benchmarking runtime pallets.
-    #[structopt(name = "benchmark", about = "Benchmark runtime pallets.")]
+    #[clap(name = "benchmark", about = "Benchmark runtime pallets.")]
     Benchmark(frame_benchmarking_cli::BenchmarkCmd),
 
     /// Try some testing command against a specified runtime state.
@@ -43,52 +43,52 @@ pub enum Subcommand {
 }
 
 /// Command for exporting the genesis state of the parachain
-#[derive(Debug, StructOpt)]
+#[derive(Debug, Parser)]
 pub struct ExportGenesisStateCommand {
     /// Output file name or stdout if unspecified.
-    #[structopt(parse(from_os_str))]
+    #[clap(parse(from_os_str))]
     pub output: Option<PathBuf>,
 
     /// Write output in binary. Default is to write in hex.
-    #[structopt(short, long)]
+    #[clap(short, long)]
     pub raw: bool,
 
     /// The name of the chain for that the genesis state should be exported.
-    #[structopt(long)]
+    #[clap(long)]
     pub chain: Option<String>,
 }
 
 /// Command for exporting the genesis wasm file.
-#[derive(Debug, StructOpt)]
+#[derive(Debug, Parser)]
 pub struct ExportGenesisWasmCommand {
     /// Output file name or stdout if unspecified.
-    #[structopt(parse(from_os_str))]
+    #[clap(parse(from_os_str))]
     pub output: Option<PathBuf>,
 
     /// Write output in binary. Default is to write in hex.
-    #[structopt(short, long)]
+    #[clap(short, long)]
     pub raw: bool,
 
     /// The name of the chain for that the genesis wasm file should be exported.
-    #[structopt(long)]
+    #[clap(long)]
     pub chain: Option<String>,
 }
 
-#[derive(Debug, StructOpt)]
-#[structopt(settings = &[
-	structopt::clap::AppSettings::GlobalVersion,
-	structopt::clap::AppSettings::ArgsNegateSubcommands,
-	structopt::clap::AppSettings::SubcommandsNegateReqs,
-])]
+#[derive(Debug, Parser)]
+#[clap(
+	propagate_version = true,
+	args_conflicts_with_subcommands = true,
+	subcommand_negates_reqs = true
+)]
 pub struct Cli {
-    #[structopt(subcommand)]
+    #[clap(subcommand)]
     pub subcommand: Option<Subcommand>,
 
-    #[structopt(flatten)]
+    #[clap(flatten)]
     pub run: cumulus_client_cli::RunCmd,
 
     /// Relay chain arguments
-    #[structopt(raw = true)]
+    #[clap(raw = true)]
     pub relay_chain_args: Vec<String>,
 }
 
@@ -119,7 +119,7 @@ impl RelayChainCli {
         Self {
             base_path,
             chain_id,
-            base: polkadot_cli::RunCmd::from_iter(relay_chain_args),
+            base: polkadot_cli::RunCmd::parse_from(relay_chain_args),
         }
     }
 }
